fp-assistant-chat/application/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
openai==0.27.2
tenacity==8.2.2
PyYAML==6.0
aioprocessing==2.0.1
pandas==2.0.3
openpyxl==3.1.2